Extended mechanic

+++ Resources/MainEditor/TemplatePanels/MenuControllerPanel.java
@@ -19,7 +19,7 @@ public class MenuControllerPanel extends JPanel {
     private int[] fontSizes;
     //panel which show changes in controller
     private MenuVisualizingPanel visualizingPanel;
-    public MenuVisualizingPanel getVisulizingPanel(){return visualizingPanel;}
+    public MenuVisualizingPanel getVisualizingPanel(){return visualizingPanel;}
     //JColorComboBox
     private JColorComboBox fontColors;
     private JColorComboBox backgroundColors;
@@ -43,6 +43,30 @@ public class MenuControllerPanel extends JPanel {
 
         //add components
         add(editComponents());
+    }
+
+    //function to avoid mixing border color with background color
+    private void avoidMixColor(){
+        if(backgroundColors.getSelectedColor().equals(Color.green)){
+            Border currSelected = BorderFactory.createLineBorder(Color.red, 5);
+            Border titledCurrSelected = BorderFactory.createTitledBorder(currSelected, "Content of : ");
+
+            //finding current selected menu
+            for(int i = 0; i < visualizingPanel.getFields().size(); i++) {
+                if(visualizingPanel.getFields().get(i).getBorder() != null)
+                    visualizingPanel.getFields().get(i).setBorder(titledCurrSelected);
+            }
+        }
+        if(backgroundColors.getSelectedColor().equals(Color.RED)){
+            Border currSelected = BorderFactory.createLineBorder(Color.GREEN, 5);
+            Border titledCurrSelected = BorderFactory.createTitledBorder(currSelected, "Content of : ");
+
+            //finding current selected menu
+            for(int i = 0; i < visualizingPanel.getFields().size(); i++) {
+                if(visualizingPanel.getFields().get(i).getBorder() != null)
+                    visualizingPanel.getFields().get(i).setBorder(titledCurrSelected);
+            }
+        }
     }
 
     //edit components
@@ -95,6 +119,13 @@ public class MenuControllerPanel extends JPanel {
         JButton deleteBtn = new JButton(" Delete ");
         deleteBtn.addActionListener(event->{
             //callback
+            //checking if we don't delete bordered menu div
+            if(visualizingPanel.getFields().get(visualizingPanel.getFields().size() - 1).getBorder() != null){
+                JOptionPane.showMessageDialog(null, "Cannot delete current editing content",
+                        "ERROR", JOptionPane.ERROR_MESSAGE);
+                return;
+            }
+
             visualizingPanel.delete();
             visualizingPanel.revalidate();
             visualizingPanel.repaint();
@@ -259,12 +290,12 @@ public class MenuControllerPanel extends JPanel {
             backgroundColors.setSelectedItem("WHITE");
             backgroundColors.addActionListener(event->{
 
+                avoidMixColor();
+
                 for(int i = 0; i < visualizingPanel.getFields().size(); i++) {
                     visualizingPanel.getFields().get(i).setBackground(backgroundColors.getSelectedColor());
                 }
                 visualizingPanel.repaint();
-                //visualizingPanel.getTextField().setBackground(backgroundColors.getSelectedColor());
-                //visualizingPanel.repaint();
             });
 
             fullComponent.add(backgroundColors);
